P100 multi-GPU fixes: FSDP offload, VAE fp16 decode, SDPA fallback
- FSDP-aware DiT offload before VAE decode (flat_param.data.cpu())
- VAE decode in float16 to halve intermediate memory
- SDPA: disable autocast + float16 cast for mem-efficient kernel on P100
- Stable sharding: scan all folders before shard split (crash-safe)
- Add run_8gpu.sh: 2x4-GPU parallel FSDP launcher
- Remove dangerous torch.set_default_dtype(float16)
- Add expandable_segments to run_8gpu.sh

+++ wan/modules/vae2_2.py
@@ -1039,7 +1039,9 @@ class Wan2_2_VAE:
         try:
             if not isinstance(zs, list):
                 raise TypeError("zs should be a list")
-            with amp.autocast(dtype=self.dtype):
+            # float16 halves intermediate memory (~976→488 MiB); output is
+            # cast to float32 by .float() below so final quality is unchanged.
+            with amp.autocast(dtype=torch.float16):
                 return [
                     self.model.decode(u.unsqueeze(0),
                                       self.scale).float().clamp_(-1,
